fix: address PR review comments on dynamic-occlusion
