--- Assets/Scripts/Debug/DisableAROcclusion.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5d2d319a8c892cb46bcd7adcec027856
