--- Assets/Scripts/Debug/EnvDepthProbe.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1a91807437fb1f14b9fe7c87729a587f
